requirements.txt
discord.py[voice]>=2.3.0
python-dotenv
requests
yt-dlp